Improve safe integer detection
Use Number.isSafeInteger on the JS side to check if value is a safe integer (which will perform all the necessary checks for both bounds and mantissa), and then convert it to `i64`.

Forward `u64`, `u128` and `i128` to this single `i64` helper since any safe integer is guaranteed to be in bounds compatible with all of these.