Deserialize only objects to structs
This removes ability to deserialise map-like data structures from JS into typed Rust structures, but that's probably semantically better anyway.

On the other hand, now we can leverage the fact that we know all the field names in advance, and instead of decoding them over and over, we encode them from the Rust side and store corresponding JS values indefinitely in a cache to use for field access.