Simplify cache hasher
Now that we're using `*const str` as a key, we don't need FNV to hash everything byte-by-byte, but instead can use the raw address of the start of the string as a hash directly. This is what I'm doing in a new custom hasher.

This allows to drop dependency on fnv, very slightly improves performance, and slightly reduces code size.